pyproject.toml
[project]
name = "project-scripts-tag"
version = "1.0.0"
description = "图片自动标注系统"
requires-python = ">=3.8.1"
authors = [
  { name = "User", email = "user@example.com" },
]
dependencies = [
  # Image processing
  "pillow>=10.0.0",

  # Web framework and API
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "python-multipart>=0.0.6",

  # Database and ORM
  "sqlalchemy>=2.0.0",

  # Configuration and utilities
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0",

  # HTTP and progress
  "requests>=2.31.0",
  "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.25.0",
  "flake8>=6.0",
  "black>=23.0.0",
  "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]